+++ src/destination/mod.rs
@@ -256,6 +256,7 @@ mod tests {
             access_key_env: None,
             secret_key_env: None,
             aws_profile: None,
+            session_token_env: None,
             allow_anonymous: false,
         };
         let dest = create_destination(&config).unwrap();
@@ -280,6 +281,7 @@ mod tests {
             access_key_env: None,
             secret_key_env: None,
             aws_profile: None,
+            session_token_env: None,
             allow_anonymous: false,
         };
         let dest = create_destination(&config).unwrap();